Returns double instead of int16
Former-commit-id: 809d07aa412854e8705028ac1fbe402615f68111

+++ +toppe/+utils/loadpfilefast.m
@@ -52,7 +52,7 @@ for icoil = 1:ncoils
         end
     end
 end
-dat = complex(datr,dati); %Combine real+imag
+dat = double(complex(datr,dati)); %Combine real+imag
 fclose(fid);
 
 return;
